Move API utilities related to game state fetching to gameStateFetchService file
