+++ frontend/src/utils/apiUtils.ts
@@ -2,95 +2,6 @@ import api from "../api.ts";
 import { MoveInfo, ParsedFEN } from "../features/gameplay/common/types/gameState.types.ts";
 import { PieceColor } from "../features/gameplay/common/types/pieces.types.ts";
 
-
-async function fetchCurrentPosition(gameId: number): Promise<ParsedFEN> {
-	let currentPosition = null;
-
-	try {
-		const response = await api.post("gameplay_api/get-current-position/", {
-			game_id: gameId,
-		});
-
-		if (response.status === 200) {
-			currentPosition = response.data;
-		}
-	} catch (error) {
-		console.error(error);
-	}
-
-	return currentPosition;
-}
-
-async function fetchTimer(gameId: number, playerColor: PieceColor) {
-	let timer = null;
-	try {
-		const response = await api.post("gameplay_api/get-player-timer/", {
-			game_id: gameId,
-			player_color: playerColor,
-		});
-
-		if (response.status === 200) {
-			timer = response.data;
-		}
-	} catch (error) {
-		console.error(error);
-	}
-
-	return timer;
-}
-
-async function fetchPositionList(gameId: number) {
-	let positionList = null;
-	try {
-		const response = await api.post("gameplay_api/get-position-list/", {
-			game_id: gameId,
-		});
-
-		if (response.status === 200) {
-			positionList = response.data;
-		}
-	} catch (error) {
-		console.error(error);
-	}
-
-	return positionList;
-}
-
-async function fetchMoveList(gameId: number) {
-	let moveList = null;
-	try {
-		const response = await api.post("gameplay_api/get-move-list/", {
-			game_id: gameId,
-		});
-
-		if (response.status === 200) {
-			moveList = response.data;
-		}
-	} catch (error) {
-		console.error(error);
-	}
-
-	return moveList;
-}
-
-async function fetchSideToMove(gameId: number) {
-	let sideToMove = null;
-
-	try {
-		const response = await api.post("/gameplay_api/get-side-to-move/", {
-			game_id: gameId,
-		});
-
-		if (response.status === 200) {
-			sideToMove = response.data;
-		}
-	} catch (error) {
-		console.log(error);
-	}
-
-	return sideToMove;
-}
-
 async function fetchLegalMoves(
 	parsedFENString: object,
 	pieceType: string,
@@ -379,10 +290,6 @@ export {
 	getIsCheckmated,
 	getIsStalemated,
 	getUsername,
-	fetchCurrentPosition,
-	fetchTimer,
-	fetchPositionList,
-	fetchMoveList,
 	createBotGame,
 	makeMoveInBotGame,
 	fetchBotGameMoveList,
@@ -390,5 +297,4 @@ export {
 	fetchCompletedGames,
 	fetchGameWinner,
 	fetchTotalCompletedGames,
-	fetchSideToMove,
 };
